Better support for gitconfig value cleanup
Git 1.7.10+ supports inclusion of external config files. By using this
feature we can make sure that gitolite maintains its git config values
in a separate file. This makes it possible to have these values
disappear when they are taken out of gitolite (instead of having to set
them as blank).